chore(tests): move helper outside __tests__ to avoid Jest empty-suite error
